Graph vis package refactor v5
