+++ src/framework/graph-visualiser/visual-graph.ts
@@ -1,6 +1,6 @@
 import { EdgeKind } from "@typedb/driver-http";
 import { MultiGraph } from "graphology";
-import { DataConstraintAny, DataVertex } from "./types";
+import { DataConstraintAny, DataVertex } from "./logical-graph";
 
 export interface VertexMetadata {
     defaultLabel: string;
